gitlab: improve pipeline definition
Clearly define workflow rules and remove unused tools like mend and ancient
dependency scanning.

Optimize the pipeline to update the dependency cache just once, but moving the
lockfile generation to the head of the pipeline and only pushing cache updates
from that job. Also build the `rsyncit.jar` only once and use the artifact to
packaging the docker image.